SED-1570-Increase test coverage
